backend/requirements.txt
flask
flask-cors
python-dotenv
pandas
numpy
xgboost
scikit-learn
requests
MetaTrader5
